+++ src/main/java/org/nrg/containers/model/command/entity/CommandEntity.java
@@ -49,6 +49,9 @@ public abstract class CommandEntity extends AbstractHibernateEntity {
     private List<CommandInputEntity> inputs;
     private List<CommandOutputEntity> outputs;
     private List<CommandWrapperEntity> commandWrapperEntities;
+    private Long reserveMemory;
+    private Long limitMemory;
+    private Double limitCpu;
 
     @Nonnull
     public static CommandEntity fromPojo(@Nonnull final Command command) {
@@ -84,6 +87,9 @@ public abstract class CommandEntity extends AbstractHibernateEntity {
         this.setCommandLine(command.commandLine());
         this.setOverrideEntrypoint(command.overrideEntrypoint());
         this.setEnvironmentVariables(command.environmentVariables());
+        this.setReserveMemory(command.reserveMemory());
+        this.setLimitMemory(command.limitMemory());
+        this.setLimitCpu(command.limitCpu());
 
         final Map<String, Command.CommandMount> mountsByName = new HashMap<>();
         for (final Command.CommandMount commandMount : command.mounts()) {
@@ -246,6 +252,30 @@ public abstract class CommandEntity extends AbstractHibernateEntity {
         this.overrideEntrypoint = overrideEntrypoint;
     }
 
+    public Long getReserveMemory() {
+        return reserveMemory;
+    }
+
+    public void setReserveMemory(Long reserveMemory) {
+        this.reserveMemory = reserveMemory;
+    }
+
+    public Long getLimitMemory() {
+        return limitMemory;
+    }
+
+    public void setLimitMemory(Long limitMemory) {
+        this.limitMemory = limitMemory;
+    }
+
+    public Double getLimitCpu() {
+        return limitCpu;
+    }
+
+    public void setLimitCpu(Double limitCpu) {
+        this.limitCpu = limitCpu;
+    }
+
     @OneToMany(mappedBy = "commandEntity", cascade = CascadeType.ALL, orphanRemoval = true)
     public List<CommandMountEntity> getMounts() {
         return mounts;
@@ -398,7 +428,10 @@ public abstract class CommandEntity extends AbstractHibernateEntity {
                 .add("environmentVariables", environmentVariables)
                 .add("inputs", inputs)
                 .add("outputs", outputs)
-                .add("xnatCommandWrappers", commandWrapperEntities);
+                .add("xnatCommandWrappers", commandWrapperEntities)
+                .add("reserveMemory", reserveMemory)
+                .add("limitMemory", limitMemory)
+                .add("limitCpu", limitCpu);
     }
 
     @Override