+++ src/main/java/org/nrg/containers/model/container/entity/ContainerEntity.java
@@ -67,7 +67,9 @@ public class ContainerEntity extends AbstractHibernateEntity {
     private List<ContainerEntityOutput> outputs;
     private List<ContainerEntityHistory> history = Lists.newArrayList();
     private List<String> logPaths;
-
+    private Long reserveMemory;
+    private Long limitMemory;
+    private Double limitCpu;
 
     public ContainerEntity() {}
 
@@ -132,6 +134,9 @@ public class ContainerEntity extends AbstractHibernateEntity {
                     }
                 }))
         );
+        this.setReserveMemory(containerPojo.reserveMemory());
+        this.setLimitMemory(containerPojo.limitMemory());
+        this.setLimitCpu(containerPojo.limitCpu());
 
         return this;
     }
@@ -299,6 +304,30 @@ public class ContainerEntity extends AbstractHibernateEntity {
 
     public void setNodeId(final String nodeId) {
         this.nodeId = nodeId;
+    }
+
+    public Long getReserveMemory() {
+        return reserveMemory;
+    }
+
+    public void setReserveMemory(final Long reserveMemory) {
+        this.reserveMemory = reserveMemory;
+    }
+
+    public Long getLimitMemory() {
+        return limitMemory;
+    }
+
+    public void setLimitMemory(final Long limitMemory) {
+        this.limitMemory = limitMemory;
+    }
+
+    public Double getLimitCpu() {
+        return limitCpu;
+    }
+
+    public void setLimitCpu(final Double limitCpu) {
+        this.limitCpu = limitCpu;
     }
 
     @ManyToOne
@@ -513,6 +542,9 @@ public class ContainerEntity extends AbstractHibernateEntity {
                 .add("outputs", outputs)
                 .add("history", history)
                 .add("logPaths", logPaths)
+                .add("reserveMemory", reserveMemory)
+                .add("limitMemory", limitMemory)
+                .add("limitCpu", limitCpu)
                 .toString();
     }
 }